requirements.txt
opencv-python
face-recognition
Pillow
numpy
cmake